refactor: #4 domain 패키지명을 model로 변경
