--- src/main/java/dev/backlog/domain/auth/model/Platform.java
@@ -1,4 +1,0 @@
-package dev.backlog.domain.auth.model;
-
-public enum Platform {
-}
