+++ src/main/java/dev/backlog/domain/comment/infrastructure/persistence/CommentRepository.java
@@ -1,6 +1,6 @@
 package dev.backlog.domain.comment.infrastructure.persistence;
 
-import dev.backlog.domain.comment.domain.Comment;
+import dev.backlog.domain.comment.model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
