+++ src/main/java/dev/backlog/domain/like/infrastructure/persistence/LikeRepository.java
@@ -1,6 +1,6 @@
 package dev.backlog.domain.like.infrastructure.persistence;
 
-import dev.backlog.domain.like.domain.Like;
+import dev.backlog.domain.like.model.Like;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
