+++ src/main/java/dev/backlog/domain/posthashtag/infrastructure/persistence/PostHashtagRepository.java
@@ -1,6 +1,6 @@
 package dev.backlog.domain.posthashtag.infrastructure.persistence;
 
-import dev.backlog.domain.posthashtag.domain.PostHashtag;
+import dev.backlog.domain.posthashtag.model.PostHashtag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostHashtagRepository extends JpaRepository<PostHashtag, Long> {
