+++ src/main/java/dev/backlog/domain/reply/infrastructure/persistence/ReplyRepository.java
@@ -1,6 +1,6 @@
 package dev.backlog.domain.reply.infrastructure.persistence;
 
-import dev.backlog.domain.reply.domain.Reply;
+import dev.backlog.domain.reply.model.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReplyRepository extends JpaRepository<Reply, Long> {
